pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reticulum-node-manager"
version = "0.1.0"
description = "Unified process manager for Reticulum mesh networking nodes with FreeDV TNC2 and Direwolf"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "W6EZE" }]
keywords = ["reticulum", "amateur-radio", "mesh-networking", "freedv", "direwolf", "packet-radio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Communications :: Ham Radio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
tui = ["textual>=0.40"]

[project.scripts]
rnm = "rnm.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/rnm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
